refactoring a bit, established intent to implement telescope picker
